refactor: 예약에 관한 예외처리를 하는 위치를 상황에 맞게 변경
